# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CheerConnect API",
    "description": "REST API for CheerConnect — a social network for the cheerleading community. Built with Next.js 16 App Router, PostgreSQL via Prisma ORM, and NextAuth.js for authentication.",
    "version": "1.0.0",
    "contact": {
      "name": "CheerConnect"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth.js JWT session cookie. Obtain by logging in via POST /api/auth/callback/credentials or OAuth flow."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Não autorizado"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
...
  "paths": {
...
    "/api/users/me/reauthenticate": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "POST /api/users/me/reauthenticate",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  }
}